pyproject.toml
[tool.poetry]
name = "dag-generator"
version = "0.1.0"
description = ""
authors = ["Nikita"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
jinja2 = "^3.1.3"
typer = "^0.9.0"


[tool.poetry.scripts]
dag-generator = "dag_generator.__main__:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
